requirements.txt
nltk
sentence-transformers
umap-learn
hdbscan
scikit-learn
bertopic
pandas
numpy
gensim
matplotlib
seaborn
sastrawi
streamlit
